requirements.txt
pymupdf 
pytesseract 
pdf2image 
easyocr 
Pillow 
torch
transformers